Rename evaluator module
