+++ evaluator/build.gradle.kts
