+++ evaluator/src/commonTest/kotlin/com/github/jcornaz/kwik/evaluator/CustomException.kt
@@ -1,4 +1,4 @@
-package com.github.jcornaz.kwik.assertions
+package com.github.jcornaz.kwik.evaluator
 
 /**
  * Dummy exception to be thrown in test and make sure that the SUT don't wrap the exceptions
