+++ evaluator/src/commonTest/kotlin/com/github/jcornaz/kwik/evaluator/ForAll4Test.kt
@@ -1,4 +1,4 @@
-package com.github.jcornaz.kwik.assertions
+package com.github.jcornaz.kwik.evaluator
 
 import com.github.jcornaz.kwik.generator.api.Generator
 import com.github.jcornaz.kwik.generator.api.withSamples
@@ -7,31 +7,43 @@ import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
-class ForAll2Test : AbstractRunnerTest() {
+class ForAll4Test : AbstractRunnerTest() {
 
     private val testGenerator1 = Generator.create { it.nextInt() }
     private val testGenerator2 = Generator.create { it.nextDouble() }
+    private val testGenerator3 = Generator.create { it.nextLong() }
+    private val testGenerator4 = Generator.create { it.nextFloat() }
 
     override fun evaluate(iterations: Int, seed: Long, invocation: () -> Boolean) {
         forAll(
             testGenerator1,
             testGenerator2,
+            testGenerator3,
+            testGenerator4,
             iterations,
             seed
-        ) { _, _ -> invocation() }
+        ) { _, _, _, _ ->
+            invocation()
+        }
     }
 
     @Test
     fun evaluateSamples() {
         val ints = mutableSetOf<Int>()
         val doubles = mutableSetOf<Double>()
+        val longs = mutableSetOf<Long>()
+        val floats = mutableSetOf<Float>()
 
         val gen1 = Generator.create { it.nextInt(0, 10) }.withSamples(42, 100)
         val gen2 = Generator.create { it.nextDouble(0.0, 10.0) }.withSamples(123.0, 678.0)
+        val gen3 = Generator.create { it.nextLong(0L, 10L) }.withSamples(-42L)
+        val gen4 = Generator.create { it.nextDouble(0.0, 10.0).toFloat() }.withSamples(-6f, 18f)
 
-        forAll(gen1, gen2) { i, d ->
+        forAll(gen1, gen2, gen3, gen4) { i, d, l, f ->
             ints += i
             doubles += d
+            longs += l
+            floats += f
             true
         }
 
@@ -39,6 +51,9 @@ class ForAll2Test : AbstractRunnerTest() {
         assertTrue(100 in ints)
         assertTrue(123.0 in doubles)
         assertTrue(123.0 in doubles)
+        assertTrue(-42L in longs)
+        assertTrue(-6f in floats)
+        assertTrue(18f in floats)
     }
 
     @Test
@@ -48,8 +63,10 @@ class ForAll2Test : AbstractRunnerTest() {
             forAll(
                 Generator.create { 42 },
                 Generator.create { -4.1 },
+                Generator.create { 100L },
+                Generator.create { "hello world" },
                 iterations = 123, seed = 78
-            ) { _, _ -> ++i < 12 }
+            ) { _, _, _, _ -> ++i < 12 }
         }
 
         assertEquals(
@@ -57,6 +74,8 @@ class ForAll2Test : AbstractRunnerTest() {
                 Property falsified after 12 tests (out of 123)
                 Argument 1: 42
                 Argument 2: -4.1
+                Argument 3: 100
+                Argument 4: hello world
                 Generation seed: 78
             """.trimIndent(),
             exception.message
@@ -67,10 +86,20 @@ class ForAll2Test : AbstractRunnerTest() {
     fun evaluateForRandomValues() {
         val valuesA = mutableSetOf<Int>()
         val valuesB = mutableSetOf<Double>()
+        val valuesC = mutableSetOf<Long>()
+        val valuesD = mutableSetOf<Float>()
 
-        forAll(testGenerator1, testGenerator2, seed = 0L) { a, b ->
+        forAll(
+            testGenerator1,
+            testGenerator2,
+            testGenerator3,
+            testGenerator4,
+            seed = 0L
+        ) { a, b, c, d ->
             valuesA += a
             valuesB += b
+            valuesC += c
+            valuesD += d
             true
         }
 
@@ -79,18 +108,30 @@ class ForAll2Test : AbstractRunnerTest() {
 
     @Test
     fun isPredictable() {
-        val pass1 = mutableListOf<Pair<Int, Double>>()
-        val pass2 = mutableListOf<Pair<Int, Double>>()
+        val pass1 = mutableListOf<Input>()
+        val pass2 = mutableListOf<Input>()
 
         val seed = 123564L
 
-        forAll(testGenerator1, testGenerator2, seed = seed) { a, b ->
-            pass1 += a to b
+        forAll(
+            testGenerator1,
+            testGenerator2,
+            testGenerator3,
+            testGenerator4,
+            seed = seed
+        ) { a, b, c, d ->
+            pass1 += Input(a, b, c, d)
             true
         }
 
-        forAll(testGenerator1, testGenerator2, seed = seed) { a, b ->
-            pass2 += a to b
+        forAll(
+            testGenerator1,
+            testGenerator2,
+            testGenerator3,
+            testGenerator4,
+            seed = seed
+        ) { a, b, c, d ->
+            pass2 += Input(a, b, c, d)
             true
         }
 
@@ -100,8 +141,10 @@ class ForAll2Test : AbstractRunnerTest() {
     @Test
     @Suppress("USELESS_IS_CHECK")
     fun canBeCalledWithoutExplicitGenerator() {
-        forAll { a: Int, b: Long ->
-            a is Int && b is Long
+        forAll { a: Int, b: Long, c: Double, d: Float ->
+            a is Int && b is Long && c is Double && d is Float
         }
     }
+
+    private data class Input(val a: Int, val b: Double, val c: Long, val d: Float)
 }
