+++ settings.gradle.kts
@@ -6,7 +6,7 @@ include("generator-api")
 include("generator-stdlib")
 include("generator-test")
 
-include("assertions")
+include("evaluator")
 
 project(":generator-api").projectDir = file("generator/api")
 project(":generator-stdlib").projectDir = file("generator/stdlib")
